requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai # For Google GenAI support
python-dotenv
pydantic-settings
pandas
tabulate
openpyxl
langchain_experimental # For experimental features
python-multipart

